requirements.txt
fastapi
uvicorn
httpx
python-dotenv
apscheduler
openai
sendgrid
pytest
pytest-asyncio
requests
sqlalchemy